Make username unique key in the database
There is still no actual authentication, but this commit is taking steps toward
that direction. Every instance of a player with same username is now the same
player.

Also restrict lightly what a username can be, so it can be printed and appear in
URLs without much issues.

TODO: This commit uses aiopg instead of asyncpg as database driver. While
asyncpg is more popular and allegedly faster library, it didn't support the
sqlalchemy_utils UUID type out of box. This needs further investigation.